v1.1: Production readiness improvements
Major changes:
- Add f16 (IEEE 754) and bf16 (Brain Float) half-precision support
- Implement Neumaier compensated summation for f32 numerical stability
- Use AVX-512 masked loads to eliminate scalar tails
- Fix uint8 L2 tail handling (signed subtraction)
- Consolidate build system: remove CMake, single Makefile pipeline
- Update CI workflow for Makefile-based builds
- Update README with new features and API

All 1508 tests pass across scalar, SSE2, AVX2, AVX-512, NEON backends.

+++ src/x86/avx2.c
@@ -339,8 +339,8 @@ uint32_t vek_l2sq_u8_avx2(const uint8_t *a, const uint8_t *b, size_t n)
     uint32_t sum_scalar = (uint32_t)_mm_cvtsi128_si32(sum);
 
     for (; i < n; i++) {
-        uint32_t diff = a[i] - b[i];
-        sum_scalar += diff * diff;
+        int32_t diff = (int32_t)a[i] - (int32_t)b[i];
+        sum_scalar += (uint32_t)(diff * diff);
     }
 
     return sum_scalar;
